requirements.txt
minio
python-dotenv
urllib3==1.26.7